Move tasks to Rakelib
After some more research, it has become apparent that the appropriate
location for custom Rake tests is actually the rakelib directory, not
the rake directory.  This change moves those files, and updates the
require statements such that none of them require require_relative any
longer.
