+++ .idea/artifacts/application_war.xml
@@ -1,0 +1,8 @@
+<component name="ArtifactManager">
+  <artifact type="war" name="application:war">
+    <output-path>$PROJECT_DIR$/test-application/target</output-path>
+    <root id="archive" name="application.war">
+      <element id="artifact" artifact-name="application:war exploded" />
+    </root>
+  </artifact>
+</component>
