+++ .idea/artifacts/application_war_exploded.xml
@@ -1,0 +1,13 @@
+<component name="ArtifactManager">
+  <artifact type="exploded-war" name="application:war exploded">
+    <output-path>$PROJECT_DIR$/test-application/target/application</output-path>
+    <root id="root">
+      <element id="javaee-facet-resources" facet="application/web/Web" />
+      <element id="directory" name="WEB-INF">
+        <element id="directory" name="classes">
+          <element id="module-output" name="application" />
+        </element>
+      </element>
+    </root>
+  </artifact>
+</component>
